move analysis code one package level up (was accidentally "analysis.analysis")
